move disable_transformers_logging to utils
